refactor(core): decompose agent module into focused submodules (#622)
- Rename streaming.rs to tool_execution.rs (#648)
- Extract message queue logic to message_queue.rs (#649)
- Extract builder methods to builder.rs
- Extract slash command handlers to commands.rs
- Extract utility helpers to utils.rs
- Reduce mod.rs from 2602 to 459 non-test lines (#650)
