+++ CHANGELOG.md
@@ -42,6 +42,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 - A2A client stream errors truncate upstream body to 256 bytes
 - Add `default_client()` HTTP helper with standard timeouts and user-agent in zeph-core and zeph-llm (#666)
 - Replace 5 production `Client::new()` calls with `default_client()` for consistent HTTP config (#667)
+- Decompose agent/mod.rs (2602→459 lines) into tool_execution, message_queue, builder, commands, and utils modules (#648, #649, #650)
 
 ### Fixed
 - False positive: "sudoku" no longer matched by "sudo" blocked pattern (word-boundary matching)
