+++ crates/zeph-core/src/agent/commands.rs
@@ -1,0 +1,198 @@
+use zeph_llm::provider::LlmProvider;
+use zeph_skills::loader::Skill;
+use zeph_skills::matcher::{SkillMatcher, SkillMatcherBackend};
+use zeph_skills::prompt::format_skills_prompt;
+use zeph_skills::registry::SkillRegistry;
+
+use crate::channel::Channel;
+use crate::config::Config;
+use crate::context::{ContextBudget, build_system_prompt};
+use zeph_tools::executor::ToolExecutor;
+
+use super::{Agent, error};
+
+impl<C: Channel, T: ToolExecutor> Agent<C, T> {
+    pub(super) async fn handle_skills_command(&mut self) -> Result<(), error::AgentError> {
+        use std::fmt::Write;
+
+        let mut output = String::from("Available skills:\n\n");
+
+        for meta in self.skill_state.registry.all_meta() {
+            let trust_info = if let Some(memory) = &self.memory_state.memory {
+                memory
+                    .sqlite()
+                    .load_skill_trust(&meta.name)
+                    .await
+                    .ok()
+                    .flatten()
+                    .map_or_else(String::new, |r| format!(" [{}]", r.trust_level))
+            } else {
+                String::new()
+            };
+            let _ = writeln!(output, "- {} — {}{trust_info}", meta.name, meta.description);
+        }
+
+        if let Some(memory) = &self.memory_state.memory {
+            match memory.sqlite().load_skill_usage().await {
+                Ok(usage) if !usage.is_empty() => {
+                    output.push_str("\nUsage statistics:\n\n");
+                    for row in &usage {
+                        let _ = writeln!(
+                            output,
+                            "- {}: {} invocations (last: {})",
+                            row.skill_name, row.invocation_count, row.last_used_at,
+                        );
+                    }
+                }
+                Ok(_) => {}
+                Err(e) => tracing::warn!("failed to load skill usage: {e:#}"),
+            }
+        }
+
+        self.channel.send(&output).await?;
+        Ok(())
+    }
+
+    pub(super) async fn handle_feedback(&mut self, input: &str) -> Result<(), error::AgentError> {
+        let Some((name, rest)) = input.split_once(' ') else {
+            self.channel
+                .send("Usage: /feedback <skill_name> <message>")
+                .await?;
+            return Ok(());
+        };
+        let (skill_name, feedback) = (name.trim(), rest.trim().trim_matches('"'));
+
+        if feedback.is_empty() {
+            self.channel
+                .send("Usage: /feedback <skill_name> <message>")
+                .await?;
+            return Ok(());
+        }
+
+        let Some(memory) = &self.memory_state.memory else {
+            self.channel.send("Memory not available.").await?;
+            return Ok(());
+        };
+
+        memory
+            .sqlite()
+            .record_skill_outcome(
+                skill_name,
+                None,
+                self.memory_state.conversation_id,
+                "user_rejection",
+                Some(feedback),
+            )
+            .await?;
+
+        if self.is_learning_enabled() {
+            self.generate_improved_skill(skill_name, feedback, "", Some(feedback))
+                .await
+                .ok();
+        }
+
+        self.channel
+            .send(&format!("Feedback recorded for \"{skill_name}\"."))
+            .await?;
+        Ok(())
+    }
+
+    pub(super) async fn reload_skills(&mut self) {
+        let new_registry = SkillRegistry::load(&self.skill_state.skill_paths);
+        if new_registry.fingerprint() == self.skill_state.registry.fingerprint() {
+            return;
+        }
+        self.skill_state.registry = new_registry;
+
+        let all_meta = self.skill_state.registry.all_meta();
+        let provider = self.provider.clone();
+        let embed_fn = |text: &str| -> zeph_skills::matcher::EmbedFuture {
+            let owned = text.to_owned();
+            let p = provider.clone();
+            Box::pin(async move { p.embed(&owned).await })
+        };
+
+        let needs_inmemory_rebuild = !self
+            .skill_state
+            .matcher
+            .as_ref()
+            .is_some_and(SkillMatcherBackend::is_qdrant);
+
+        if needs_inmemory_rebuild {
+            self.skill_state.matcher = SkillMatcher::new(&all_meta, embed_fn)
+                .await
+                .map(SkillMatcherBackend::InMemory);
+        } else if let Some(ref mut backend) = self.skill_state.matcher
+            && let Err(e) = backend
+                .sync(&all_meta, &self.skill_state.embedding_model, embed_fn)
+                .await
+        {
+            tracing::warn!("failed to sync skill embeddings: {e:#}");
+        }
+
+        let all_skills: Vec<Skill> = self
+            .skill_state
+            .registry
+            .all_meta()
+            .iter()
+            .filter_map(|m| self.skill_state.registry.get_skill(&m.name).ok())
+            .collect();
+        let trust_map = self.build_skill_trust_map().await;
+        let skills_prompt = format_skills_prompt(&all_skills, std::env::consts::OS, &trust_map);
+        self.skill_state
+            .last_skills_prompt
+            .clone_from(&skills_prompt);
+        let system_prompt = build_system_prompt(&skills_prompt, None, None, false);
+        if let Some(msg) = self.messages.first_mut() {
+            msg.content = system_prompt;
+        }
+
+        tracing::info!(
+            "reloaded {} skill(s)",
+            self.skill_state.registry.all_meta().len()
+        );
+    }
+
+    pub(super) fn reload_config(&mut self) {
+        let Some(ref path) = self.config_path else {
+            return;
+        };
+        let config = match Config::load(path) {
+            Ok(c) => c,
+            Err(e) => {
+                tracing::warn!("config reload failed: {e:#}");
+                return;
+            }
+        };
+
+        self.runtime.security = config.security;
+        self.runtime.timeouts = config.timeouts;
+        self.memory_state.history_limit = config.memory.history_limit;
+        self.memory_state.recall_limit = config.memory.semantic.recall_limit;
+        self.memory_state.summarization_threshold = config.memory.summarization_threshold;
+        self.skill_state.max_active_skills = config.skills.max_active_skills;
+        self.skill_state.disambiguation_threshold = config.skills.disambiguation_threshold;
+
+        if config.memory.context_budget_tokens > 0 {
+            self.context_state.budget = Some(ContextBudget::new(
+                config.memory.context_budget_tokens,
+                0.20,
+            ));
+        } else {
+            self.context_state.budget = None;
+        }
+        self.context_state.compaction_threshold = config.memory.compaction_threshold;
+        self.context_state.compaction_preserve_tail = config.memory.compaction_preserve_tail;
+        self.context_state.prune_protect_tokens = config.memory.prune_protect_tokens;
+        self.memory_state.cross_session_score_threshold =
+            config.memory.cross_session_score_threshold;
+
+        #[cfg(feature = "index")]
+        {
+            self.index.repo_map_ttl =
+                std::time::Duration::from_secs(config.index.repo_map_ttl_secs);
+        }
+
+        tracing::info!("config reloaded");
+    }
+}
